Testing with the Player Character
"have the player run slower and slower the farther ahead they are from
the ideal X position, and run faster to catch up to the ideal X
position. The player should lose momentum most of their momentum if they
hit an enemy, because the dash afterwards can give them momentum if they
wanted it. Also I put the player in my level but it's really hard to
click on enemies and dash to them for some reason"