Enemy Models Added & Adjusted
Considering reworking the enemy script to spawn and manage its 3d model
prefab at the start of the game in order to let us make changes to the
3d model more easily. Bug where the player is killed by the enemy
they're dashing to is still present. Player still gets stuck on walls.
Drag feature needs to be edited to be more user-friendly.